Added seconds to dateformat
Alternatively, add the ability to customize the default date format through the Admin Panel plugin (long and short date formats as well). They are now based on the list defined in the blueprint. However, there is the option to edit 'system.pages.dateformat.default' directly in system.yaml, but this is not comfortable and the value is only temporary until another is selected and saved. I prefer these values: 'd.m.Y H: i: s' and 'd-m-Y H: i: s'. I will describe this in more detail in the core and thre admin issues.